Replaced resolution option with ambient volume option

+++ _Project/BGMVolumeController.h
@@ -3,6 +3,7 @@
 #include "kitten\K_Component.h"
 #include "puppy\Text\TextBox.h"
 
+// I really should have just made an options_menu controller - Callum
 class BGMVolumeController : public kitten::K_Component
 {
 private:
@@ -15,6 +16,5 @@ public:
 	~BGMVolumeController();
 
 	void changeBGMVolume(int p_amount);
-
-	float getCurrentBGMVolume() const { return m_volume; };
+	int getCurrentBGMVolume() const { return m_volume; };
 };
